Organize snail directory structure
