--- src/snail/Makefile
@@ -1,44 +1,0 @@
-AR = ar
-CXX = clang++
-CXX_FLAGS = -Wall -Wextra -g -O0 -MMD -DDEBUG -std=c++14 -I/usr/local/include
-BIN_DIR = bin
-SRC_DIR = .
-SOURCES = $(wildcard $(SRC_DIR)/*.cpp) $(wildcard $(SRC_DIR)/*/*.cpp)
-OBJECTS = $(foreach i, $(SOURCES), $(BIN_DIR)/$(basename $(i)).o)
-LIB = ./snail.a
-
-
-MKDIR := mkdir
-
-
-.PHONY: build clean rebuild
-
-
-all: build
-
-
-
-build: $(BIN_DIR) $(LIB)
-
-
-$(BIN_DIR):
-	$(MKDIR) -p $(BIN_DIR)/detail
-
-
-$(LIB): $(OBJECTS)
-	$(AR) r $@ $^
-
-
-$(BIN_DIR)/%.o: $(SRC_DIR)/%.cpp
-	$(CXX) $(CXX_FLAGS) -c $< -o $@
-
-
-clean:
-	-@$(RM) -f $(PROGRAM)
-	-@$(RM) -f $(BIN_DIR)/*.d $(BIN_DIR)/*.o
-
-
-rebuild: clean build
-
-
--include $(BIN_DIR)/*.d
