+++ src/snail/backends/headless/application.cpp
@@ -1,6 +1,7 @@
+#include "../../application.hpp"
 #include <algorithm>
 #include <map>
-#include "../input.hpp"
+#include "../../input.hpp"
 
 
 namespace elona
