+++ src/snail/backends/headless/detail.cpp
@@ -1,3 +1,7 @@
+#include "../../detail/sdl.hpp"
+
+
+
 namespace elona
 {
 namespace snail
