+++ src/snail/backends/headless/window.cpp
@@ -1,3 +1,7 @@
+#include "../../window.hpp"
+
+
+
 namespace elona
 {
 namespace snail
