+++ src/snail/backends/sdl/filedialog.cpp
@@ -1,7 +1,8 @@
+#include "../../filedialog.hpp"
 #include <cstdlib>
 #include <type_traits>
-#include "../../thirdparty/nfd/nfd.h"
-#include "../../util/scope_guard.hpp"
+#include "../../../thirdparty/nfd/nfd.h"
+#include "../../../util/scope_guard.hpp"
