+++ src/snail/backends/sdl/window.cpp
@@ -1,6 +1,4 @@
-#ifdef ANDROID
-#include "../window.hpp"
-#endif
+#include "../../window.hpp"
