--- src/snail/touch_input.cpp
@@ -1,9 +1,0 @@
-#include "touch_input.hpp"
-
-#if defined(SNAIL_RENDERER_HEADLESS)
-#include "touch_input/headless.cpp"
-#elif defined(SNAIL_RENDERER_SDL)
-#include "touch_input/sdl.cpp"
-#else
-#error Unsupported renderer
-#endif
